# api/docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.LevelCreate:
    properties:
      language_id:
        type: integer
      name:
        type: string
    type: object
  models.LevelResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      language_id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.LevelUpdate:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
  title: Asrlan-Monolithic
paths:
...
  /v1/level:
    post:
      consumes:
      - application/json
      description: This Api for creating a new level
      parameters:
      - description: LevelCreate Model
        in: body
        name: LevelCreate
        required: true
        schema:
          $ref: '#/definitions/models.LevelCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LevelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Create Level
      tags:
      - levels
    put:
      consumes:
      - application/json
      description: This Api for updating level
      parameters:
      - description: Update LevelUpdate Model
        in: body
        name: LevelUpdate
        required: true
        schema:
          $ref: '#/definitions/models.LevelUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LevelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update Level
      tags:
      - levels
  /v1/level/{id}:
    delete:
      consumes:
      - application/json
      description: This Api for deleting level
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete Level
      tags:
      - levels
    get:
      consumes:
      - application/json
      description: This Api for get level
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LevelResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get Level
      tags:
      - levels
  /v1/levels:
    get:
      consumes:
      - application/json
      description: This Api for get all levels
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LevelResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: ListLevels
      tags:
      - levels
  /v1/levelsforcourse:
    get:
      consumes:
      - application/json
      description: This Api for get all levels
      parameters:
      - description: UserId
        in: query
        name: user_id
        required: true
        type: string
      - description: LanguageId
        in: query
        name: language_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LevelResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: ListLevels
      tags:
      - levels
  /v1/levelsforregister:
    get:
      consumes:
      - application/json
      description: This Api for get all levels
      parameters:
      - description: LanguageId
        in: query
        name: language_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LevelResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: ListLevels
      tags:
      - levels
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"